:ok_hand: :truck: Moved layouts components
