--- back-office/src/App/App.css
@@ -1,17 +1,0 @@
-.App {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
-
-.App > main {
-  display: flex;
-}
-
-.App > main, .App > main > * {
-  flex: 1;
-}
-
-.App > header, .App > footer {
-  z-index: 100;
-}
