+++ back-office/src/layouts/AppLayout/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { default } from './AppLayout'
