+++ back-office/src/layouts/index.js
@@ -1,0 +1,3 @@
+// @flow
+export { default as AppLayout } from './AppLayout'
+export { default as MainScreen } from './MainScreen'
